+++ docs/content/components/content/header/simple-header.demo.tsx
@@ -2,7 +2,7 @@ import { Header, Headline, Text } from '@marigold/components';
 
 export default () => (
   <Header>
-    <Headline level="3">Header!</Headline>
+    <Headline level={3}>Header!</Headline>
     <Text>awesome</Text>
   </Header>
 );