+++ docs/content/components/layout/center/text-center.demo.tsx
@@ -2,7 +2,7 @@ import { Button, Center, Headline, Stack, Text } from '@marigold/components';
 
 export default () => (
   <Stack space={4}>
-    <Headline level="2">Star Wars - The Empire Strikes Back</Headline>
+    <Headline level={3}>Star Wars - The Empire Strikes Back</Headline>
     <Text>
       It is a dark time for the Rebellion. Although the Death Star has been
       destroyed, Imperial troops have driven the Rebel forces from their hidden