+++ packages/components/src/Headline/Headline.tsx
@@ -1,4 +1,5 @@
-import { ReactNode } from 'react';
+import { Heading } from 'react-aria-components';
+import type RAC from 'react-aria-components';
 
 import {
   TextAlignProp,
@@ -9,31 +10,24 @@ import {
   useClassNames,
   useTheme,
 } from '@marigold/system';
-import { HtmlProps } from '@marigold/types';
 
-// Props
-// ---------------
+type RemovedProps = 'className' | 'level';
 export interface HeadlineProps
-  extends Omit<HtmlProps<'h1'>, 'className'>,
+  extends Omit<RAC.HeadingProps, RemovedProps>,
     TextAlignProp {
-  children?: ReactNode;
-  level?: '1' | '2' | '3' | '4' | '5' | '6';
   color?: string;
   variant?: string;
   size?: string;
-  as?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6';
+  level?: '1' | '2' | '3' | '4' | '5' | '6' | RAC.HeadingProps['level'];
 }
 
-// Component
-// ---------------
-export const Headline = ({
-  children,
+const _Headline = ({
   variant,
   size,
+  children,
   align = 'left',
   color,
-  level = '1',
-  as = `h${level}`,
+  level = 1,
   ...props
 }: HeadlineProps) => {
   const theme = useTheme();
@@ -43,10 +37,9 @@ export const Headline = ({
     size: size ?? `level-${level}`,
   });
 
-  const Component = as;
-
   return (
-    <Component
+    <Heading
+      level={Number(level)}
       {...props}
       className={cn(classNames, 'text-[--color]', textAlign[align])}
       style={createVar({
@@ -57,6 +50,8 @@ export const Headline = ({
       })}
     >
       {children}
-    </Component>
+    </Heading>
   );
 };
+
+export { _Headline as Headline };